% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within.R
\name{within}
\alias{within}
\title{Select observations within a given range}
\usage{
within(x, a, b)
}
\arguments{
\item{x}{A vector}

\item{a}{Lower limit}

\item{b}{Upper limit}
}
\value{
A logical vector of the same length as \code{x}
}
\description{
Select observations within a given range
}
\examples{
within(1:10, 2, 5)

}

